+++ modules/conductingMaterial/conductingMaterial.C
@@ -49,33 +49,28 @@ Foam::solvers::conductingMaterial::conductingMaterial
     solver(mesh),
     electroBase(mesh)
 {
-    // Non-source conductingMaterial regions are considered passive.
-    // Initialize only source regions.
-    //if (electro.isSource())
-    //{
-        label PotERefCell = 0;
-        scalar PotERefValue = 0.0;
+    label PotERefCell = 0;
+    scalar PotERefValue = 0.0;
+    setRefCell
+    (
+        electro.PotE(),
+        pimple.dict(),
+        PotERefCell,
+        PotERefValue
+    );
+    mesh.schemes().setFluxRequired(electro.PotE().name());
+
+    if (electro.isComplex())
+    {
         setRefCell
         (
-            electro.PotE(),
+            electro.PotE(true),
             pimple.dict(),
             PotERefCell,
             PotERefValue
         );
-        mesh.schemes().setFluxRequired(electro.PotE().name());
-
-        if (electro.isComplex())
-        {
-            setRefCell
-            (
-                electro.PotE(true),
-                pimple.dict(),
-                PotERefCell,
-                PotERefValue
-            );
-            mesh.schemes().setFluxRequired(electro.PotE(true).name());
-        }
-    //}
+        mesh.schemes().setFluxRequired(electro.PotE(true).name());
+    }
 }
 
 
@@ -121,6 +116,9 @@ void Foam::solvers::conductingMaterial::pressureCorrector()
 
 void Foam::solvers::conductingMaterial::postCorrector()
 {
+    // Presently set as passive.
+    // Uncomment to enable current calculation.
+
     // Solve only for source regions
     if (electro.correctElectromagnetics())
     {